refactor: js파일 모듈화
