+++ constants/index.js
@@ -1,0 +1,1 @@
+export const NEWS_COUNT = 24;
