+++ scripts/rolling-banner.js
@@ -1,0 +1,64 @@
+const $containerNewsBar = document.querySelector(".container-news-bar");
+
+const createRollingBannerList = ($banner, data) => {
+  // 첫번째 요소를 마지막에 copy
+  const list = [...data, data[0]];
+
+  $banner.innerHTML = list.reduce((acc, curr) => {
+    return acc + `<li><a href="/">${curr}</a></li>`;
+  }, "");
+};
+
+export const startRollingBanner = (headlineData) => {
+  const [left, right] = [headlineData.slice(0, 5), headlineData.slice(5)];
+  const dataLength = left.length;
+
+  const $banners = $containerNewsBar.querySelectorAll(
+    ".container-news-bar_window > ul"
+  );
+
+  createRollingBannerList($banners[0], left);
+  createRollingBannerList($banners[1], right);
+
+  $banners.forEach(($banner, idx) => {
+    let cnt = 0;
+    let interval;
+
+    const autoRolling = () => {
+      cnt += 1;
+      $banner.style.transitionDuration = "500ms";
+      $banner.style.transform = `translateY(-${14 * cnt}px)`;
+
+      if (cnt >= dataLength) {
+        clearInterval(interval);
+        setTimeout(() => {
+          cnt = 0;
+          $banner.style.transitionDuration = "0ms";
+          $banner.style.transform = "translateY(0)";
+          autoPlay();
+        }, 500);
+      }
+    };
+
+    const autoPlay = () => {
+      interval = setInterval(autoRolling, 5 * 1000);
+    };
+
+    if (idx === 1) {
+      setTimeout(autoPlay, 1000);
+    } else {
+      autoPlay();
+    }
+
+    const handleArticleMouseOver = () => {
+      clearInterval(interval);
+    };
+
+    const handleArticleMouseOut = () => {
+      autoPlay();
+    };
+
+    $banner.addEventListener("mouseover", handleArticleMouseOver);
+    $banner.addEventListener("mouseout", handleArticleMouseOut);
+  });
+};
