+++ utils/sort.js
@@ -1,0 +1,3 @@
+export const shuffleData = (data) => {
+  return data.sort(() => Math.random() - 0.5);
+};
